pyproject.toml
[tool.poetry]
name = "file-server"
version = "0.1.0"
description = "File server"
authors = ["lunerien <lukasz.kala6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
django-auth-ldap = "^4.6.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
